Final Changes made to project. All code is commented. Will revisit to expand beyond the requirements.

+++ src/main/java/DAO/SocialMediaDAO.java
@@ -14,6 +14,8 @@ import Util.ConnectionUtil;
 /**
  * Database Access Object create sql statements and executes them to the database.
  * Methods have names based on the action being taken on the database (e.x select, drop, update, insert, etc.)
+ * Turns strings into prepared statements. One reason to to this is to prevent sql injections by checking the input via prepared statement
+ * The other is so it is easy to add variable with ?
  */
 public class SocialMediaDAO {
     //put connection here so it doesn't declare it in each method scope.
@@ -25,12 +27,22 @@ public class SocialMediaDAO {
      * @return returns account with account_id field on success, null on failure
      */
     public Account selectAccount(Account account) {
+        //Catches unexpected errors and returns null
         try {
+            //sql command string
             String sql = "SELECT * FROM account WHERE username = ? AND password = ?";
+
+            //turns string into prepared statement
             PreparedStatement ps = this.connection.prepareStatement(sql);
+
+            //Adds values for ? fields
             ps.setString(1, account.getUsername());
             ps.setString(2, account.getPassword());
+
+            //Gets Result
             ResultSet rs = ps.executeQuery();
+
+            //If there is a result then create object if it fails return null
             while(rs.next()) {
                 Account returnAccount = new Account(rs.getInt("account_id"),
                     rs.getString("username"),
@@ -52,20 +64,33 @@ public class SocialMediaDAO {
      * @return returns the account object with the assigned id now. Returns null on error
      * */
     public Account insertAccount(Account account) {
+
+        //Catches unexpected errors and returns null
         try {
+            //Sql string
             String sql = "INSERT INTO account (username, password) VALUES (?, ?);";
+
+            //Turns into prepared statement
             PreparedStatement ps = this.connection.prepareStatement(sql, Statement.RETURN_GENERATED_KEYS);
+
+            //Adds values for ? fields
             ps.setString(1, account.getUsername());
             ps.setString(2, account.getPassword());
+
+            //Inserts into accounts table 
             ps.executeUpdate();
+
+            //Gets generated id for object return
             ResultSet rs = ps.getGeneratedKeys();
+
+            //If rs exists then set id to account and return otherwise return null
             if(rs.next()){
                 int account_id = (int) rs.getLong(1);
                 account.setAccount_id(account_id);
                 return account;
             }
             
-        } catch (SQLException e) {
+        } catch (Exception e) {
             System.out.println(e.getMessage());
         }
         return null;
@@ -73,10 +98,12 @@ public class SocialMediaDAO {
 
     /**
      * Deletes account record from database not in requirements
+     * TODO work on once initial project is submitted
      * @param account
      * @return
      */
     public Account deleteAccount(Account account) {
+        //Catches unexpected errors and returns null
         try {
             String sql = "DELETE FROM account WHERE account_id = ?";
             PreparedStatement ps = this.connection.prepareStatement(sql);
@@ -95,16 +122,31 @@ public class SocialMediaDAO {
 
     //
     public Message insertMessage(Message message) {
+        //Catches unexpected errors and returns null
         try {
+            //sql string
             String sql = "INSERT INTO message (posted_by, message_text, time_posted_epoch) VALUES (?, ?, ?)";
-            PreparedStatement ps = this.connection.prepareStatement(sql);
+
+            //creates prepared statement that can return the generated key
+            PreparedStatement ps = this.connection.prepareStatement(sql, Statement.RETURN_GENERATED_KEYS);
+
+            //Add values to sql command
             ps.setInt(1, message.getPosted_by());
             ps.setString(2, message.getMessage_text());
             ps.setLong(3, message.getTime_posted_epoch());
+
+            //inserts message
             int updated = ps.executeUpdate();
+
+            //If updated is greater than 0 then it was inserted
             if(updated > 0){
-                System.out.println("Message Inserted");
-                return message;
+                //sets id to message and returns it, null on fail or not inserted
+                ResultSet rs = ps.getGeneratedKeys();
+                if(rs.next()){
+                    int message_id = (int) rs.getLong(1);
+                    message.setMessage_id(message_id);
+                    return message;
+                }
             }
         } catch (Exception e) {
             System.out.println(e.getMessage());
@@ -112,10 +154,20 @@ public class SocialMediaDAO {
         return null;
     }
     public Message selectMessage(int message_id) {
+        //Catches unexpected errors and returns null
         try {
+            //sql string
             String sql = "SELECT * FROM message WHERE message_id = ?";
+
+            //turns it into a prepared statement
             PreparedStatement ps = this.connection.prepareStatement(sql);
+
+            //Adds value for id
+            ps.setInt(1, message_id);
+
+            //gets result from select statement
             ResultSet rs = ps.executeQuery();
+            //If there is a result, build and return object otherwise return null
             while(rs.next()){
                 Message mesage = new Message(
                     rs.getInt("message_id"),
@@ -131,10 +183,18 @@ public class SocialMediaDAO {
     }
     public List<Message> selectAllMessages() {
         List<Message> messages = new ArrayList<>();
+        //Catches unexpected errors and returns null
         try {
+            //sql statement
             String sql = "SELECT * FROM message";
+
+            //turns it into prepared statment
             PreparedStatement ps = this.connection.prepareStatement(sql);
+
+            //Gets reuslts
             ResultSet rs = ps.executeQuery();
+
+            //If results exist, itterate over all of them and build and return list of messages, null on failure 
             while(rs.next()){
                 Message mesage = new Message(
                     rs.getInt("message_id"),
@@ -151,11 +211,21 @@ public class SocialMediaDAO {
     }
     public List<Message> selectAllMessagesFromAccountId(int account_id) {
         List<Message> messages = new ArrayList<>();
+        //Catches unexpected errors and returns null
         try {
+            //sql statement
             String sql = "SELECT * FROM message WHERE posted_by = ?";
+
+            //turns it into prepared statment
             PreparedStatement ps = this.connection.prepareStatement(sql);
+
+            //Adds value for id
             ps.setInt(1, account_id);
+
+            //Gets reuslts
             ResultSet rs = ps.executeQuery();
+
+            //If results exist, itterate over all of them and build and return list of messages, null on failure 
             while(rs.next()){
                 Message mesage = new Message(
                     rs.getInt("message_id"),
@@ -164,6 +234,7 @@ public class SocialMediaDAO {
                     rs.getLong("time_posted_epoch"));
                 messages.add(mesage);
             }
+            System.out.println(messages.toString());
             return messages;
         } catch (Exception e) {
             System.out.println(e.getMessage());
@@ -171,12 +242,27 @@ public class SocialMediaDAO {
         return null;
     }
     public Message deleteMessage(int message_id){
+        //Catches unexpected errors and returns null
         try {
+            //First checks if message exists from select message if it does we need this to return deleted message
             Message message = selectMessage(message_id);
+            if(message == null) {
+                return null;
+            }
+
+            //sql statement
             String sql = "DELETE FROM message WHERE message_id = ?;";
+
+            //turns it into prepared statment
             PreparedStatement ps = this.connection.prepareStatement(sql);
+
+            //sets field value
             ps.setInt(1, message_id);
+
+            //execute update and get returned int
             int deleted = ps.executeUpdate();
+
+            //If int returned is more than 0 then a message was deleted
             if(deleted > 0){
                 System.out.println("Message Deleted");
                 return message;
@@ -187,6 +273,9 @@ public class SocialMediaDAO {
         }
         return null;
     }
+    //TODO Delete all messages by id
+    //Helps deal with table reference if deleteing something referenced without using cascade
+    //Auditing would also require a deleted messages table for admin use
     /**
      * 
      * @param message_id id value of the message provided from the pathParam
@@ -194,13 +283,22 @@ public class SocialMediaDAO {
      * @return
      */
     public Message updateMessage(int message_id, String message_text){
+        //Catches unexpected errors and returns null
         try {
+            //sql statement
             String sql = "UPDATE message SET message_text = ? WHERE message_id = ?";
+
+            //turns statement into prepared statement
             PreparedStatement ps = this.connection.prepareStatement(sql);
+
+            //set field values
             ps.setString(1, message_text);
+            ps.setInt(2, message_id);
+
+            //gets returned int on update
             int updated = ps.executeUpdate();
+            //If updated is more than 0, then  return message otherwise return null
             if(updated > 0){
-                System.out.println("Message Updated");
                 return selectMessage(message_id);
             }
         } catch (Exception e) {
